backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.7
python-decouple==3.8 
django-cors-headers==4.0.0
Pillow==10.0.0